Refactor ClientExtensions class
