+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/controller/ChannelControllerImpl.kt
@@ -59,9 +59,7 @@ import io.getstream.chat.android.livedata.ChatDomainImpl
 import io.getstream.chat.android.livedata.controller.helper.MessageHelper
 import io.getstream.chat.android.livedata.extensions.NEVER
 import io.getstream.chat.android.livedata.extensions.addMyReaction
-import io.getstream.chat.android.livedata.extensions.isImageMimetype
 import io.getstream.chat.android.livedata.extensions.isPermanent
-import io.getstream.chat.android.livedata.extensions.isVideoMimetype
 import io.getstream.chat.android.livedata.extensions.removeMyReaction
 import io.getstream.chat.android.livedata.extensions.shouldIncrementUnreadCount
 import io.getstream.chat.android.livedata.extensions.wasCreatedAfter
@@ -1503,6 +1501,10 @@ internal class ChannelControllerImpl(
     internal fun replyMessage(repliedMessage: Message?) {
         _repliedMessage.value = repliedMessage
     }
+
+    private fun String?.isImageMimetype() = this?.contains("image") ?: false
+
+    private fun String?.isVideoMimetype() = this?.contains("video") ?: false
 
     companion object {
         private const val TYPE_IMAGE = "image"
