--- stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/extensions/LiveDataExtensions.kt
@@ -1,10 +1,0 @@
-package io.getstream.chat.android.livedata.extensions
-
-import androidx.lifecycle.MutableLiveData
-import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
-import kotlinx.coroutines.withContext
-
-internal suspend fun <T : Any?> MutableLiveData<T>.setOnUi(newValue: T) =
-    withContext(DispatcherProvider.Main) {
-        value = newValue
-    }
