+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/extensions/Message.kt
@@ -1,0 +1,32 @@
+package io.getstream.chat.android.livedata.extensions
+
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.models.Reaction
+import io.getstream.chat.android.client.models.User
+import java.util.Date
+
+internal val NEVER = Date(0)
+
+internal fun Message.wasCreatedAfterOrAt(date: Date?): Boolean {
+    return createdAt ?: createdLocallyAt ?: NEVER >= date
+}
+
+internal fun Message.wasCreatedAfter(date: Date?): Boolean {
+    return createdAt ?: createdLocallyAt ?: NEVER > date
+}
+
+internal fun Message.wasCreatedBefore(date: Date?): Boolean {
+    return createdAt ?: createdLocallyAt ?: NEVER < date
+}
+
+internal fun Message.wasCreatedBeforeOrAt(date: Date?): Boolean {
+    return createdAt ?: createdLocallyAt ?: NEVER <= date
+}
+
+internal fun Message.users(): List<User> {
+    return latestReactions.mapNotNull(Reaction::user) + user + (replyTo?.users().orEmpty())
+}
+
+internal fun Message.shouldIncrementUnreadCount(currentUserId: String): Boolean {
+    return user.id != currentUserId && !silent && !shadowed
+}
